Move API response types to a separate file
Tidy up the types into a separate file. These are shared between
several files and the verbose definitions harm readability of the
functional code.
