+++ src/content-block/api-client.ts
@@ -1,43 +1,9 @@
-/**
- * Represents an embed code response from the API.
- */
-export type EmbedCodePreview =
-  | { html: string; valid: true; error: null }
-  | { html: null; valid: false; error: Error };
-
-export enum BlockType {
-  Pension = "Pension",
-  Contact = "Contact",
-  TimePeriod = "Time Period",
-}
-
-/**
- * The organisation that owns a content block.
- */
-export interface ContentBlockOrganisation {
-  name: string;
-  content_id: string;
-}
-
-/**
- * A single content block as returned by the blocks list endpoint.
- */
-export interface ContentBlock {
-  title: string;
-  block_type: BlockType;
-  organisation: ContentBlockOrganisation;
-  state: string;
-  embed_code: string;
-  formats: string[];
-}
-
-/**
- * The response shape of the blocks list endpoint.
- */
-export interface BlocksResponse {
-  results: ContentBlock[];
-}
-
+import {
+  BlockType,
+  ContentBlock,
+  BlocksResponse,
+  EmbedCodePreview,
+} from "../@types";
 import { isValidEmbedCode } from "./regex.ts";
 
 const supportedBlockTypes = new Set<string>(Object.values(BlockType));
